requirements.txt
# Core ML/AI dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
diffusers>=0.18.0
accelerate>=0.20.0

# Image processing
Pillow>=9.5.0
opencv-python>=4.7.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0
datasets>=2.13.0
jsonlines>=3.1.0

# Math and symbolic computation
sympy>=1.12
scipy>=1.10.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0

# Web scraping (optional)
requests>=2.31.0
beautifulsoup4>=4.12.0
praw>=7.7.0

# GUI/Demo (optional)
streamlit>=1.24.0
gradio>=3.35.0

# Development tools
jupyter>=1.0.0
ipywidgets>=8.0.0
black>=23.0.0
pytest>=7.4.0
